+++ Matterhook.NET/Webhooks/Discourse/Shared.cs
@@ -2,8 +2,8 @@
 {
     public class Actions_Summary
     {
-        public int id { get; set; }
-        public int count { get; set; }
+        public long id { get; set; }
+        public long count { get; set; }
         public bool hidden { get; set; }
         public bool can_act { get; set; }
     }
